refactor: rearrange packages
