+++ didcomm_messaging/v1/__init__.py
@@ -1,0 +1,10 @@
+"""DIDComm v1 Interfaces.
+
+These components are intended to provide a similar structure to the DIDComm v2
+interfaces provided by this library. While the community is transitioning from
+DIDComm v1 to v2, having a consistent interface to interact with will help
+implementers to support both versions until the transition is complete.
+
+It is expected that a future version of this library will eventually remove
+these interfaces.
+"""
