+++ pyproject.toml
@@ -31,6 +31,10 @@ legacy = [
     "msgpack>=1.0.8",
     "PyNaCl>=1.5.0",
 ]
+nacl = [
+    "msgpack>=1.0.8",
+    "PyNaCl>=1.5.0",
+]
 
 [build-system]
 requires = ["pdm-backend"]
